requirements.txt
pyTelegramBotApi==3.7.1
pymongo==3.10.1
mysql-connector-python==8.0.20
